--- src/pages/Home.jsx
@@ -1,13 +1,0 @@
-// Components
-import Features from "../components/home/Features";
-import { LandingBottom, LandingTop } from "../components/home/LandingBanner";
-
-export default function Home() {
-  return (
-    <>
-      <LandingTop />
-      <Features />
-      <LandingBottom />
-    </>
-  );
-}
