+++ src/pages/Home.tsx
@@ -1,0 +1,12 @@
+import Features from "@/components/home/Features";
+import { LandingBottom, LandingTop } from "@/components/home/LandingBanner";
+
+export default function Home() {
+  return (
+    <>
+      <LandingTop />
+      <Features />
+      <LandingBottom />
+    </>
+  );
+}
